feat: Add workflow vars for template name and entrypoint
Add two new global workflow variables.

Issue #6772 asks for the template name, but we cannot always know what
this is (simplest unknowable is kubectl apply of a workflow with a
generateName:) So provide that by deduction. See the tests added in
this commit for what we're aiming for there.

Also provide the variable mainEndpoint, which is set to the name of
the invoked entrypoint, which is controllable, knowable, and could be
used in the same way, and seems like a useful thing to have anyway.

Addresses #6772.

Signed-off-by: Alan Clucas <alan@clucas.org>