fix(tests): cron e2e tests run in parallel
testify/suite breaks parallel cron testing which was introduced in #2118

The breakage probably happened in #2874

testify know this https://github.com/stretchr/testify/issues/187 and
will not be fixing it.

The fix here is inspired by @Jerska in https://github.com/stretchr/testify/issues/187#issuecomment-1380086748

It has a potential race but we won't be affected by it as all the
tests take time to run.

The tests are designed around running starting on a minute boundary
for reliabilty, and as we add more tests without this fix it
* increases overall reliability
* test-to-test runtime varies, changing the start time relative to a
minute boundary of a test, increasing uncertainty in how many
instances of a cron will run.

Tested locally by running the suite

Signed-off-by: Alan Clucas <alan@clucas.org>